+++ docs/source/entry/HorizontalFlex.js
@@ -3,14 +3,13 @@
  * @author vivaxy
  */
 
-
 import React, { Component } from 'react';
 
 import connect from '../library/connect';
 import render from '../library/render';
 import setTitle from '../library/setTitle';
 
-import { VerticalFlexBox, VerticalFlexItem } from 'react-pianist/VerticalFlex';
+import { HorizontalFlexBox, HorizontalFlexItem } from 'react-pianist/HorizontalFlex';
 import colors from 'react-pianist/colors';
 
 @connect(state => ({}), {})
@@ -18,26 +17,31 @@ class HorizontalFlex extends Component {
 
     render () {
 
-        setTitle(`VerticalFlexBox`);
+        setTitle(`HorizontalFlex`);
 
-        return <VerticalFlexBox style={{
+        return <HorizontalFlexBox style={{
             width: '100%',
             height: '100%',
             position: 'absolute',
         }}>
-            <VerticalFlexItem
+            <HorizontalFlexItem
                 flex={1}
                 style={{
                     backgroundColor: colors.TURQUOISE,
+                    color: colors.WHITE,
+                    lineHeight: '50px',
+
                 }}
-            >this block auto expand</VerticalFlexItem>
-            <VerticalFlexItem
-                height={50}
+            >this block auto expand</HorizontalFlexItem>
+            <HorizontalFlexItem
+                flex={1}
                 style={{
                     backgroundColor: colors.EMERALD,
+                    color: colors.WHITE,
+                    lineHeight: '50px',
                 }}
-            >this block will not expand</VerticalFlexItem>
-        </VerticalFlexBox>
+            >this block auto expand too</HorizontalFlexItem>
+        </HorizontalFlexBox>
     }
 
 }